requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
psycopg2-binary==2.9.9
asgiref==3.7.0
